입력에 대한 validate 로직을 Lotto로 이동
