+++ src/main/kotlin/Lotto.kt
@@ -1,3 +1,7 @@
+import LottoCard.Companion.LOTTO_LAST_NUMBER
+import LottoCard.Companion.LOTTO_NUMBER_CNT
+import LottoCard.Companion.LOTTO_START_NUMBER
+
 fun main() {
     val lotto = Lotto()
     val price = lotto.validatePrice(inputPrice())
@@ -5,7 +9,8 @@ fun main() {
     val lottoCards = LottoCards(lottoCnt)
     printLottoCards(lottoCards)
 
-    val beforeWeekLottoCard = LottoCard(inputLottoNumber())
+    val numbers = lotto.validateLottoCard(inputLottoNumber())
+    val beforeWeekLottoCard = LottoCard(numbers)
     val statistic = lotto.getStatistic(lottoCards, beforeWeekLottoCard)
     val yieldRate = lotto.getYieldRate(statistic, price)
     printResult(statistic, yieldRate)
@@ -23,6 +28,18 @@ class Lotto {
         }
     }
 
+    fun validateLottoCard(numberLine: String?): List<Int> {
+        require(!numberLine.isNullOrBlank()) { "로또 번호를 반드시 입력해야 합니다." }
+
+        val strNumbers = numberLine.replace(WHITESPACE_REGEX, EMPTY_STRING).split(",")
+        require(strNumbers.size == LOTTO_NUMBER_CNT) { "로또 번호는 6개입니다." }
+
+        val numbers = strNumbers.map { it.parseInt() }
+        require(numbers.none { it < LOTTO_START_NUMBER || it > LOTTO_LAST_NUMBER }) { "입력된 숫자가 로또 번호의 범위 밖입니다." }
+
+        return numbers
+    }
+
     fun validatePrice(strPrice: String?): Int {
         require(!strPrice.isNullOrBlank()) { "구입금액을 반드시 입력해야합니다." }
 
@@ -38,5 +55,7 @@ class Lotto {
 
     companion object {
         private const val LOTTO_PRICE = 1000
+        private val WHITESPACE_REGEX = Regex("\\s")
+        private const val EMPTY_STRING = ""
     }
 }
